backend/requirements.txt
# Core
python-dotenv==1.0.0
Pillow==10.4.0

# API
flask==3.0.0
flask-cors==4.0.0

# AI/ML
groq==0.9.0
opencv-python-headless==4.8.1.78

# Database
supabase==2.0.3

# Utils
requests==2.31.0
python-multipart==0.0.6